Add controller-section.js file and add view function
Add controller-section.js file
it will control the events in the sections div in the sidebar in the future, right
now it gets the names of the sections from the section model and gets the div of the
DOM from the section Div in the sidebar and it sends them to view to render the different
sections.

controller.js
Now it initialize controllerSection and calls the initialize function of controller
Section.

view.js
New function renderSections that gets divDom and the section names and sends them
to sidebar.js function sidebarSectionTemplate

sidebar.js
new function sidebarSectionTemplate gets sectionNames and the div of the sections
and builds the HTML with a button for every sectionName.