Controller now save current section or project
it created two setters and two getters
one for current section and one for current project.
as currentproject saves the projectID and current Sections
saves the currentSection string ex. 'Home'

the setter for sections also resets currentProject to null and
the setter for project resets currentSection to '' because they
can not be loaded at the same time.

Setters are passed as callback to controller section and controller
project to actualize when loading a new page.