Fix some bugs
It tried to load a Page projects when clicking on the project
Div but without clicking in any specific project.

Add a little white outline for the tasks so if we put a grey
project so a grey background for the task it differenciates
from the main background color.

Add a border also for the projects color in the sidebar.

Change the Mock Data