added debounce to color change event
Now when changing the color, the tasks will also change their color, this is executed
with a debounce function to not trigger too many loadings that could slow down the page.

Also change the date functions file and filter functions file to an utils folder with
the debounce function file.