...
When changing date for example being in the month section and change the date out of
the month with details open the details are still open even when the task is not in
the screen anymore added a functionality to automatically close the details.

BUT !!!!!!
the functionality is commented because i don't know which one is better by design,
for example if the date is cahnged from details by mistake and the details are open
you can change it back and put the task again in the month, if the details are closed
you can't change the date again. That could frustate the user.